refactor(vc-processors): examples
