--- venus-worker/vc-processors-demos/customize-proc/README.md
@@ -1,2 +1,0 @@
-# customize-proc
-This demo shows how to implement a customize processor for a builtin task.
