--- venus-worker/vc-processors-demos/sqr-consumer/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "sqr-consumer"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-anyhow = "1"
-
-[dependencies.vc-processors]
-path = "../../vc-processors"
-default-features = false
