+++ venus-worker/vc-processors/examples/sqr-consumer.rs
@@ -1,3 +1,20 @@
+//! This demo shows how to implment an ext consumer for some specific `Task`.
+//!
+//! You can run the compiled binary, enter something like `{"id": 1, "task": 15}` to the input,
+//! and then see the result from stdout.
+//!
+//! ```
+//! $ cargo run --example sqr-consumer
+//! 2022-06-01T07:41:34.114535Z  INFO sqr_consumer: start sqr consumer
+//! pow processor ready
+//! 2022-06-01T07:41:34.114678Z  INFO sub{name=pow pid=10951}: vc_processors::core::ext::consumer: processor ready
+//! {"id": 1, "task": 15}
+//! {"id":1,"err_msg":null,"output":225}
+//! ```
+//!
+//! You can use any number you like to replace `15`, and see what happens.
+//!
+
 use anyhow::{Context, Result};
 use serde::{Deserialize, Serialize};
 use tracing::info;
@@ -9,7 +26,7 @@ use vc_processors::core::{ext::run_consumer, Processor, Task};
 struct Num(pub u32);
 
 impl Task for Num {
-    const STAGE: &'static str = "pow";
+    const STAGE: &'static str = "sqr";
     type Output = Num;
 }
